Out-of-bounds straggler checks run always
Moved certain straggler checks (particularly those for being outside of the world or off-navmesh) so that they are run even if an NPC is in-combat. Previously NPCs could avoid being relocated if they were actively shooting a player (which in the case of out-of-bounds enemies is a case where we ESPECIALLY want to move them).