Vort adjustments / fixes
Fixed the code meant to stop vorts from using their shield ability when no one is present. A typo was stopping it from excluding itself.

Added visuals to indicate when Vortigaunts have a charge ready, so that players can interact with this mechanic more effectively (either abusing it for their own gain, or picking them off to prevent them using it).