Stop sentinel teleporting out of bounds
Areas that don't connect to the main map, like the skybox, are no longer valid teleport locations.

This happened because the fall-back algorithm for if the player's zone was unknown chose a random zone. Now it only uses the main zone (guaranteeing you're in the same area as the objectives / enemies).