Re-organised prefab definitions
Prefabs are now defined in an auto-included types file like missions, classes, factions, announcers, etc.

I've split this into a file for natural prefabs and a file for objectives. Individually these then have groupings for related prefabs (e.g. all of the "ambient" natural prefabs are in one group, all of the terminals/hackables are in another).