Scaling required sweepers for nukes
The sweeper requirement for activating a nuke now scales down over time, starting at 5 minutes, and eventually reaching 0.

This is to address an unintended side-effect of the system, that being making it drastically harder to activate nukes in radiation once one has been detonated.

This should largely preserve the intended effect (preventing detonations that are the fault of a singular player, rather than the team, which usually happens when people activate them immediately on round start). While preventing it from stalling out rounds for too long.