Exclude clips from validArea traces
Excludes player and NPC clips from the traces used in validAreas.

Hitting a playerclip would previously trigger nodraw checks, making the check think it's out of bounds or underneath the map. This is a relatively niche case, as it's not typical for there to be giant playerclips over the top of large sections of the map, but the creator of lakeview did exactly that for some reason, and it might cause weird behaviour in other cases.