Incomplete zombie spreading
Has issues. Currently doesn't work as intended. Uncomment code in zombieSlowThink to test.

This update reads in the connections for the nodegraph, and adds an A* implementation for navigating them.

It also standardises A* implementations (mostly) so that we have less repeated code.

Navigation behaviour similar to that of Airgraph code is here, but it's currently incomplete. In theory npcs using this navigation code should be forced to spread out / use different paths to reach a target.